try to catch up to nuts

+++ database/gkvdb/level.go
@@ -12,10 +12,11 @@ import (
 var _ Database = (*levelDB)(nil)
 
 type LevelConfig struct {
-	Home string
+	Home   string
+	Tables []string
 }
 
-func DefaultLevelConfig(home string) *LevelConfig {
+func DefaultLevelConfig(home string, tables []string) *LevelConfig {
 	return &LevelConfig{
 		Home: home,
 	}
@@ -23,6 +24,8 @@ func DefaultLevelConfig(home string) *LevelConfig {
 
 type levelDB struct {
 	db *leveldb.DB
+
+	tables map[string]struct{}
 
 	cfg *LevelConfig
 }
@@ -32,7 +35,14 @@ func NewLevelDB(cfg *LevelConfig) (Database, error) {
 		return nil, ErrInvalidConfig
 	}
 	bdb := &levelDB{
-		cfg: cfg,
+		cfg:    cfg,
+		tables: make(map[string]struct{}, len(cfg.Tables)),
+	}
+	for _, v := range cfg.Tables {
+		if _, ok := bdb.tables[v]; ok {
+			return nil, ErrDuplicateTable
+		}
+		bdb.tables[v] = struct{}{}
 	}
 
 	return bdb, nil
@@ -58,14 +68,23 @@ func (b *levelDB) Close(_ context.Context) error {
 }
 
 func (b *levelDB) Del(_ context.Context, table string, key []byte) error {
+	if _, ok := b.tables[table]; !ok {
+		return ErrTableNotFound
+	}
 	return b.db.Delete(NewCompositeKey(table, key), nil)
 }
 
 func (b *levelDB) Has(_ context.Context, table string, key []byte) (bool, error) {
+	if _, ok := b.tables[table]; !ok {
+		return false, ErrTableNotFound
+	}
 	return b.db.Has(NewCompositeKey(table, key), nil)
 }
 
 func (b *levelDB) Get(_ context.Context, table string, key []byte) ([]byte, error) {
+	if _, ok := b.tables[table]; !ok {
+		return nil, ErrTableNotFound
+	}
 	value, err := b.db.Get(NewCompositeKey(table, key), nil)
 	if err != nil {
 		if errors.Is(err, leveldb.ErrNotFound) {
@@ -77,6 +96,12 @@ func (b *levelDB) Get(_ context.Context, table string, key []byte) ([]byte, erro
 }
 
 func (b *levelDB) Put(_ context.Context, table string, key, value []byte) error {
+	if key == nil {
+		return ErrInvalidKey
+	}
+	if _, ok := b.tables[table]; !ok {
+		return ErrTableNotFound
+	}
 	return b.db.Put(NewCompositeKey(table, key), value, nil)
 }
 
@@ -86,6 +111,7 @@ func (b *levelDB) Begin(_ context.Context, write bool) (Transaction, error) {
 		return nil, xerr(err)
 	}
 	return &levelTX{
+		db: b,
 		tx: tx,
 	}, nil
 }
@@ -99,10 +125,16 @@ func (b *levelDB) Update(ctx context.Context, callback func(ctx context.Context,
 }
 
 func (b *levelDB) NewIterator(ctx context.Context, table string) (Iterator, error) {
+	if _, ok := b.tables[table]; !ok {
+		return nil, ErrTableNotFound
+	}
 	return nil, fmt.Errorf("not yet")
 }
 
 func (b *levelDB) NewRange(ctx context.Context, table string, start, end []byte) (Range, error) {
+	if _, ok := b.tables[table]; !ok {
+		return nil, ErrTableNotFound
+	}
 	return nil, fmt.Errorf("not yet")
 }
 
@@ -113,24 +145,40 @@ func (b *levelDB) NewBatch(ctx context.Context) (Batch, error) {
 // Transactions
 
 type levelTX struct {
+	db *levelDB
 	tx *leveldb.Transaction
 }
 
 func (tx *levelTX) Del(ctx context.Context, table string, key []byte) error {
+	if _, ok := tx.db.tables[table]; !ok {
+		return ErrTableNotFound
+	}
 	return xerr(tx.tx.Delete(NewCompositeKey(table, key), nil))
 }
 
 func (tx *levelTX) Has(ctx context.Context, table string, key []byte) (bool, error) {
+	if _, ok := tx.db.tables[table]; !ok {
+		return false, ErrTableNotFound
+	}
 	has, err := tx.tx.Has(NewCompositeKey(table, key), nil)
 	return has, xerr(err)
 }
 
 func (tx *levelTX) Get(ctx context.Context, table string, key []byte) ([]byte, error) {
+	if _, ok := tx.db.tables[table]; !ok {
+		return nil, ErrTableNotFound
+	}
 	value, err := tx.tx.Get(NewCompositeKey(table, key), nil)
 	return value, xerr(err)
 }
 
 func (tx *levelTX) Put(ctx context.Context, table string, key []byte, value []byte) error {
+	if key == nil {
+		return ErrInvalidKey
+	}
+	if _, ok := tx.db.tables[table]; !ok {
+		return ErrTableNotFound
+	}
 	return xerr(tx.tx.Put(NewCompositeKey(table, key), value, nil))
 }
